requirements.txt
pytest
allure-pytest
selenium
openpyxl
pytest-html